css_parse: move CursorToSourceCursorSink to its own file
This was getting a little big and deserved being split out. This also makes it easier to find using a file fuzzy finder,
as the file name matches the struct name.
