+++ crates/css_parse/src/lib.rs
@@ -288,6 +288,7 @@ mod cursor_interleave_sink;
 mod cursor_ordered_sink;
 mod cursor_overlay_sink;
 mod cursor_pretty_write_sink;
+mod cursor_to_source_cursor_sink;
 mod cursor_write_sink;
 mod diagnostics;
 mod feature;
@@ -312,6 +313,7 @@ pub use cursor_interleave_sink::*;
 pub use cursor_ordered_sink::*;
 pub use cursor_overlay_sink::*;
 pub use cursor_pretty_write_sink::*;
+pub use cursor_to_source_cursor_sink::*;
 pub use cursor_write_sink::*;
 pub use diagnostics::*;
 pub use feature::*;
